feat: step 1 compiled
Description
You just started sorting out the existing database of the "Easy Rider" bus company. As you take the first look at the data, you realize that it's not going to be easy.

Sometimes numbers are missing from where they should definitely be. You also noticed that sometimes there are too many or too few characters. Fortunately, there is documentation to help you sort out this mess. However, this documentation is not a hundred percent complete: part of it was torn away when your colleague spilled coffee on it. Let's see what we can make out.

Here are the documents that you have: documentation and diagram of the bus lines.

Objectives
The string containing the data in JSON format is passed to standard input.
Check that the data types match.
Check that the required fields are filled in.
Display the information about the number of found errors in total and in each field. Keep in mind that there might be no errors at all.
The output should have the same formatting as shown in the example.
No need to worry about the format (name validation) now. Let's at first just make sure that the fields have the right type and all required ones are filled.
If you can't find the necessary information in the stage description, it can probably be found in the attached documentation.

Note that the type Char is present among the data types.